requirements.txt
# Core Dependencies
pandas==2.1.4
numpy==1.26.2

# LLM & Embeddings
google-generativeai>=0.8.0
sentence-transformers>=2.3.0
huggingface-hub>=0.20.0

# Vector Store
faiss-cpu==1.7.4

# UI
streamlit==1.29.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
tabulate==0.9.0
